pyproject.toml
[project]
name = "toy-localization-agent"
version = "1.0.0"
description = "Multi-agent toy localization assistant"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "langchain==0.1.0",
    "langchain-community==0.0.20",
    "langchain-openai==0.0.6",
    "chromadb==0.4.22",
    "tiktoken>=0.5.2,<1.0.0",
    "requests>=2.31.0,<3.0.0",
    "pillow>=10.0.0,<13.0.0",
    "openai>=1.10.0,<2",
    "python-dotenv>=1.0.0,<2",
    "dashscope>=1.25.12",
    "fastapi>=0.110.0,<1.0.0",
    "uvicorn>=0.27.0,<1.0.0",
]

[project.scripts]
toy-localization = "main:main"
toy-localization-web = "webapp:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agents", "knowledge", "web"]
py-modules = [
    "config",
    "workflow",
    "webapp",
    "main",
    "formatting",
    "schemas",
    "market_normalizer",
    "feature_heuristics",
    "language_utils",
    "verification",
]

[tool.setuptools.package-data]
knowledge = ["data/*.txt"]
web = ["index.html", "assets/*.css", "assets/*.js"]
